requirements.txt
psycopg2-binary
pymongo
pandas
kagglehub
python-dotenv
fastapi
uvicorn[standard]
pydantic
